requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
pypdf
python-docx
html2text
python-dotenv
tiktoken
sqlalchemy
psycopg2-binary
python-multipart
email
